backend/requirements.txt
Django>=5.0,<6
djangorestframework>=3.15,<4
djangorestframework-simplejwt>=5.3,<6
psycopg2-binary>=2.9,<3
dj-database-url>=2.2,<3
django-cors-headers>=4.3,<5
python-dotenv>=1.0,<2
gunicorn>=22,<24
whitenoise>=6.6,<7
